pyproject.toml
[project]
name = "ai-employee"
version = "0.3.0"
description = "Personal AI Employee - Gold Tier"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "apscheduler>=3.10.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "tweepy>=4.14.0",
    "tenacity>=8.2.0",
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.0",
]

[project.scripts]
file-watcher = "watchers.filesystem_watcher:main"
gmail-watcher = "watchers.gmail_watcher:main"
orchestrator = "watchers.orchestrator:main"
